Refactor main daemon
* This will now only run once a week, fetching all historic
price data for all stocks that we support. This will ensure
that if a new stock enters the index, we'll fetch all of the
data for it in the next cycle

* This is very ineffecient at the moment because we're fetching
records for millions of items that we already have. This also
takes like 5 days to run, so it's far from an ideal solution.

* This should be fine for the short term though