Update fetchHeatmapDataForDate method to rely on WMH stock
I had a problem where the historic price updates are interrupted
every time the server processes are stopped, which is basically
every time I deploy something new. Because of this, the first stock
would likely have data, but the last would not. As I was only checking
the first stock in the list for whether it had data, and using this
to assume all of the subsequent stocks would, this was a problem.

To solve this, just use a stock you can count on existing later in the
list.